fix old tests mirroring the mongo adapter files
